requirements.txt
asgiref==3.7.2
astroid==3.1.0
dill==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytz==2024.1
sqlparse==0.4.4
tomlkit==0.12.3
